Moves conformance-specific test fns out of accounts-db
